git-svn-id: svn://127.0.0.1/simple-stratum-proxy@1074 1c29cedf-3501-9543-bf67-87c9abb7e30a

+++ src/main/java/strat/mining/stratum/proxy/json/MiningSetDifficultyNotification.java
@@ -1,3 +1,21 @@
+/**
+ * stratum-proxy is a proxy supporting the crypto-currency stratum pool mining
+ * protocol.
+ * Copyright (C) 2014  Stratehm (stratehm@hotmail.com)
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with multipool-stats-backend. If not, see <http://www.gnu.org/licenses/>.
+ */
 package strat.mining.stratum.proxy.json;
 
 import java.util.ArrayList;
@@ -12,7 +30,7 @@ public class MiningSetDifficultyNotification extends JsonRpcNotification {
 	public static final String METHOD_NAME = "mining.set_difficulty";
 
 	@JsonIgnore
-	private Integer difficulty;
+	private Double difficulty;
 
 	public MiningSetDifficultyNotification() {
 		super(METHOD_NAME);
@@ -22,11 +40,11 @@ public class MiningSetDifficultyNotification extends JsonRpcNotification {
 		super(notification);
 	}
 
-	public Integer getDifficulty() {
+	public Double getDifficulty() {
 		return difficulty;
 	}
 
-	public void setDifficulty(Integer difficulty) {
+	public void setDifficulty(Double difficulty) {
 		this.difficulty = difficulty;
 	}
 
@@ -44,7 +62,7 @@ public class MiningSetDifficultyNotification extends JsonRpcNotification {
 	public void setParams(List<Object> params) {
 		super.setParams(params);
 		if (params != null) {
-			difficulty = ((Number) getParamsObjectAtIndex(0)).intValue();
+			difficulty = ((Number) getParamsObjectAtIndex(0)).doubleValue();
 		}
 	}
