+++ src/main/java/strat/mining/stratum/proxy/model/Share.java
@@ -1,9 +1,27 @@
+/**
+ * stratum-proxy is a proxy supporting the crypto-currency stratum pool mining
+ * protocol.
+ * Copyright (C) 2014  Stratehm (stratehm@hotmail.com)
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with multipool-stats-backend. If not, see <http://www.gnu.org/licenses/>.
+ */
 package strat.mining.stratum.proxy.model;
 
 public class Share {
 
 	private long time;
-	private int difficulty;
+	private double difficulty;
 
 	public long getTime() {
 		return time;
@@ -13,11 +31,11 @@ public class Share {
 		this.time = time;
 	}
 
-	public int getDifficulty() {
+	public double getDifficulty() {
 		return difficulty;
 	}
 
-	public void setDifficulty(int difficulty) {
+	public void setDifficulty(double difficulty) {
 		this.difficulty = difficulty;
 	}
