+++ src/main/java/strat/mining/stratum/proxy/rest/dto/PoolDetailsDTO.java
@@ -1,3 +1,21 @@
+/**
+ * stratum-proxy is a proxy supporting the crypto-currency stratum pool mining
+ * protocol.
+ * Copyright (C) 2014  Stratehm (stratehm@hotmail.com)
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with multipool-stats-backend. If not, see <http://www.gnu.org/licenses/>.
+ */
 package strat.mining.stratum.proxy.rest.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
@@ -22,8 +40,8 @@ public class PoolDetailsDTO {
 
 	private Integer priority;
 
-	private Long acceptedDifficulty;
-	private Long rejectedDifficulty;
+	private Double acceptedDifficulty;
+	private Double rejectedDifficulty;
 
 	private Boolean isExtranonceSubscribeEnabled;
 
@@ -131,19 +149,19 @@ public class PoolDetailsDTO {
 		this.isActiveSince = isActiveSince;
 	}
 
-	public Long getAcceptedDifficulty() {
+	public Double getAcceptedDifficulty() {
 		return acceptedDifficulty;
 	}
 
-	public void setAcceptedDifficulty(Long acceptedDifficulty) {
+	public void setAcceptedDifficulty(Double acceptedDifficulty) {
 		this.acceptedDifficulty = acceptedDifficulty;
 	}
 
-	public Long getRejectedDifficulty() {
+	public Double getRejectedDifficulty() {
 		return rejectedDifficulty;
 	}
 
-	public void setRejectedDifficulty(Long rejectedDifficulty) {
+	public void setRejectedDifficulty(Double rejectedDifficulty) {
 		this.rejectedDifficulty = rejectedDifficulty;
 	}
